Add user role to `/whoami` response
