+++ registry/registry_ui_test.go
@@ -16,7 +16,6 @@ import (
 	"github.com/lestrrat-go/jwx/v2/jwk"
 	"github.com/pelicanplatform/pelican/test_utils"
 	"github.com/pkg/errors"
-	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -245,87 +244,6 @@ func TestGetNamespaceJWKS(t *testing.T) {
 			if tc.expectedCode == http.StatusOK {
 				assert.Equal(t, tc.expectedData, w.Body.String())
 			}
-		})
-	}
-}
-
-func TestAdminAuthHandler(t *testing.T) {
-	// Initialize Gin and set it to test mode
-	gin.SetMode(gin.TestMode)
-
-	// Define test cases
-	testCases := []struct {
-		name          string
-		setupUserFunc func(*gin.Context) // Function to setup user and admin list
-		expectedCode  int                // Expected HTTP status code
-		expectedError string             // Expected error message
-	}{
-		{
-			name: "user-not-logged-in",
-			setupUserFunc: func(ctx *gin.Context) {
-				viper.Set("Registry.AdminUsers", []string{"admin1", "admin2"})
-				ctx.Set("User", "")
-			},
-			expectedCode:  http.StatusUnauthorized,
-			expectedError: "Login required to view this page",
-		},
-		{
-			name: "general-admin-access",
-			setupUserFunc: func(ctx *gin.Context) {
-				viper.Set("Registry.AdminUsers", []string{})
-				ctx.Set("User", "admin")
-			},
-			expectedCode: http.StatusOK,
-		},
-		{
-			name: "specific-admin-user-access",
-			setupUserFunc: func(ctx *gin.Context) {
-				viper.Set("Registry.AdminUsers", []string{"admin1", "admin2"})
-				ctx.Set("User", "admin1")
-			},
-			expectedCode: http.StatusOK,
-		},
-		{
-			name: "non-admin-user-access",
-			setupUserFunc: func(ctx *gin.Context) {
-				viper.Set("Registry.AdminUsers", []string{"admin1", "admin2"})
-				ctx.Set("User", "user")
-			},
-			expectedCode:  http.StatusForbidden,
-			expectedError: "You don't have permission to perform this action",
-		},
-		{
-			name: "admin-list-empty",
-			setupUserFunc: func(ctx *gin.Context) {
-				viper.Set("Registry.AdminUsers", []string{})
-				ctx.Set("User", "user")
-			},
-			expectedCode:  http.StatusForbidden,
-			expectedError: "You don't have permission to perform this action",
-		},
-		{
-			name: "admin-list-multiple-users",
-			setupUserFunc: func(ctx *gin.Context) {
-				viper.Set("Registry.AdminUsers", []string{"admin1", "admin2", "admin3"})
-				ctx.Set("User", "admin2")
-			},
-			expectedCode: http.StatusOK,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			w := httptest.NewRecorder()
-			ctx, _ := gin.CreateTestContext(w)
-			tc.setupUserFunc(ctx)
-
-			adminAuthHandler(ctx)
-
-			assert.Equal(t, tc.expectedCode, w.Code)
-			if tc.expectedError != "" {
-				assert.Contains(t, w.Body.String(), tc.expectedError)
-			}
-			viper.Reset()
 		})
 	}
 }
